garimind_mvp_starter/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
alembic==1.13.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
greenlet==3.1.1
aiosqlite==0.20.0

google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
msal==1.30.0
requests==2.32.3

openai==1.51.0
sse-starlette==2.1.0
